Added ability to use AppData directory
Added ability to use AppData directory as opposed to only using the
current working directory. (Saving the config file(s) on the fly when
they reside in ProgramFiles didn't work unless the program was run as
administrator.)

This functionality was NOT added to the MTGCardRecognizer, but changes
needed to be made because of changes made to Core to support using
AppData in BeadMaker.